pyproject.toml
[tool.poetry]
name = "simple-blog"
version = "0.1.0"
description = "It is a simple blog as a test for dzencode"
authors = ["volodymyr-v-konoval <volodymyr.v.konoval@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.5.0"
python-dotenv = "^1.0.1"
django-simple-captcha = "^0.6.2"
psycopg2-binary = "^2.9.10"
channels = "^4.2.0"
channels-redis = "^4.2.1"
daphne = "^4.1.2"
elasticsearch-dsl = "^8.17.1"
django-elasticsearch-dsl = "^8.0"
celery = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
